standup/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-standup-server"
version = "0.1.0"
description = "MCP server for generating standup reports from GitHub activity"
authors = [
  { name = "Brett", email = "20871604+brettinternet@users.noreply.github.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "mcp>=1.0.0",
  "python-dateutil>=2.8.0",
  "pydantic>=2.0.0",
  "rich>=13.0.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "pytest-mock>=3.10.0",
  "freezegun>=1.2.0",
  "ruff>=0.1.0",
]

[project.scripts]
mcp-standup-server = "mcp_server_standup.server:main_sync"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
